pyproject.toml
[tool.poetry]
name = "homework-django"
version = "0.1.0"
description = ""
authors = ["Romancov Maksim <romancov_dzr@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1"
psycopg2-binary = "^2.9.9"
pillow = "^10.4.0"
ipython = "^8.27.0"
python-slugify = {extras = ["unidecode"], version = "^8.0.4"}
slugify = "^0.0.1"
psycopg2 = "^2.9.9"
requests = "^2.32.3"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^8.0.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
